fixed db_session

+++ backend/model/profile.py
@@ -1,5 +1,5 @@
 import sqlalchemy
-from sqlalchemy import orm, Column, VARCHAR, DOUBLE, ForeignKey
+from sqlalchemy import orm, Column, VARCHAR, ForeignKey, INTEGER
 from .db_session import Base
 
 
@@ -9,12 +9,12 @@ class Profile(Base):
     profile_id = Column(VARCHAR(200), primary_key=True, nullable=False)
     username = Column(VARCHAR(200), ForeignKey('User.username'),
                       nullable=False)
-    balance = Column(DOUBLE, default=0)
-    investment = Column(DOUBLE, default=0)
-    dollars = Column(DOUBLE, default=0)
-    euro = Column(DOUBLE, default=0)
-    yuan = Column(DOUBLE, default=0)
-    bitcoin = Column(DOUBLE, default=0)
+    balance = Column(INTEGER, default=0)
+    investment = Column(INTEGER, default=0)
+    dollars = Column(INTEGER, default=0)
+    euro = Column(INTEGER, default=0)
+    yuan = Column(INTEGER, default=0)
+    bitcoin = Column(INTEGER, default=0)
 
     user = orm.relationship('User', back_populates='profile')
